buildscripts/requirements.txt
pyyaml == 3.11
typing == 3.5
unittest-xml-reporting == 2.1.0
